More hrtim refactor
